adopt cmd/sub-cmd cli to allow for future features
