+++ go.mod
@@ -4,14 +4,24 @@ go 1.12
 
 require (
 	github.com/coreos/go-systemd v0.0.0-20181031085051-9002847aa142
-	github.com/gin-contrib/sse v0.0.0-20170109093832-22d885f9ecc7
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/gin-contrib/sse v0.0.0-20170109093832-22d885f9ecc7 // indirect
 	github.com/gin-gonic/gin v0.0.0-20180417025440-6ad7b9c9d382
-	github.com/godbus/dbus v0.0.0-20181101234600-2ff6f7ffd60f
-	github.com/golang/protobuf v0.0.0-20180328163153-e09c5db29600
-	github.com/json-iterator/go v0.0.0-20180125064802-358cfc39298c
-	github.com/mattn/go-isatty v0.0.4
-	github.com/ugorji/go v0.0.0-20180407103000-f3cacc17c85e
-	golang.org/x/sys v0.0.0-20180126165840-ff2a66f350ce
-	gopkg.in/go-playground/validator.v8 v8.18.2
-	gopkg.in/yaml.v2 v2.0.0
+	github.com/godbus/dbus v0.0.0-20181101234600-2ff6f7ffd60f // indirect
+	github.com/golang/protobuf v0.0.0-20180328163153-e09c5db29600 // indirect
+	github.com/google/gofuzz v0.0.0-20170612174753-24818f796faf // indirect
+	github.com/inconshreveable/mousetrap v1.0.0 // indirect
+	github.com/json-iterator/go v0.0.0-20180125064802-358cfc39298c // indirect
+	github.com/mattn/go-isatty v0.0.4 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/spf13/cobra v0.0.3
+	github.com/spf13/pflag v1.0.3 // indirect
+	github.com/stretchr/testify v1.2.2 // indirect
+	github.com/ugorji/go/codec v0.0.0-20190204201341-e444a5086c43 // indirect
+	golang.org/x/net v0.0.0-20190301231341-16b79f2e4e95 // indirect
+	golang.org/x/sync v0.0.0-20190227155943-e225da77a7e6 // indirect
+	golang.org/x/sys v0.0.0-20181205085412-a5c9d58dba9a // indirect
+	gopkg.in/go-playground/assert.v1 v1.2.1 // indirect
+	gopkg.in/go-playground/validator.v8 v8.18.2 // indirect
+	gopkg.in/yaml.v2 v2.2.2 // indirect
 )
